ScalafixImplSuite does not test cli internals, just the interfaces API
Spread the tests in other test suites
